Refactor dataloaders module
