+++ scvi/dataloaders/__init__.py
@@ -1,12 +1,10 @@
-from ._ann_dataloader import AnnDataLoader
-from ._anntorchdataset import AnnTorchDataset
-from ._concat_dataloader import ConcatDataLoader
+from ._dataloaders import AnnDataLoader, ConcatDataLoader, SemiSupervisedDataLoader
+from ._datasets import AnnTorchDataset
 from ._datasplitters import (
     DataSplitter,
     DeviceBackedDataSplitter,
     SemiSupervisedDataSplitter,
 )
-from ._semi_dataloader import SemiSupervisedDataLoader
 
 __all__ = [
     "AnnDataLoader",
